+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   responses ([rhbz#2097731])
 - Set HTTP headers to prevent caching everything except static files
   ([rhbz#2097733])
+- Set HTTP headers to prevent sending referrer ([rhbz#2097732])
 
 ### Added
 - Add support for fence\_mpath to `pcs stonith update-scsi-devices` command
@@ -45,6 +46,7 @@
 [rhbz#2068457]: https://bugzilla.redhat.com/show_bug.cgi?id=2068457
 [rhbz#2095695]: https://bugzilla.redhat.com/show_bug.cgi?id=2095695
 [rhbz#2097731]: https://bugzilla.redhat.com/show_bug.cgi?id=2097731
+[rhbz#2097732]: https://bugzilla.redhat.com/show_bug.cgi?id=2097732
 [rhbz#2097733]: https://bugzilla.redhat.com/show_bug.cgi?id=2097733
 [huntr#220307]: https://huntr.dev/bounties/7aa921fc-a568-4fd8-96f4-7cd826246aa5/